refactor: split CommandExt
